Rename jugglingTrickSlice to jugglingTricksSlice
